chapter_14_debug/requirements.txt
django-rest-framework
django-log-request-id
